chore: add functional migration to our internal Gatsby Plugins
